+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- context: `Context::recv()` and `Context::try_recv()` panics if called again after returning `None`.
+
 ## [0.1.27] - 2021-09-27
 ### Fixed
 - telemeter: remove duplicate actor_group/actor_key labels.